Moving command of database from user servises to database file
