+++ src/app/users/routers.py
@@ -1,6 +1,8 @@
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
 
+import database
+
 from . import schemas, services
 
 
@@ -8,9 +10,7 @@ router = APIRouter()
 
 
 @router.post("/", response_model=schemas.User, tags=["users"])
-def create_user(
-    user: schemas.UserCreate, db: Session = Depends(services.get_db)
-):
+def create_user(user: schemas.UserCreate, db: Session = Depends(database.get_db)):
     db_user = services.get_user_by_email(db, email=user.email)
     if db_user:
         raise HTTPException(status_code=400, detail="Email already registered")
@@ -18,15 +18,13 @@ def create_user(
 
 
 @router.get("/", response_model=list[schemas.User], tags=["users"])
-def read_users(
-    skip: int = 0, limit: int = 100, db: Session = Depends(services.get_db)
-):
+def read_users(skip: int = 0, limit: int = 100, db: Session = Depends(database.get_db)):
     users = services.get_users(db, skip=skip, limit=limit)
     return users
 
 
 @router.get("/{user_id}", response_model=schemas.User, tags=["users"])
-def read_user(user_id: int, db: Session = Depends(services.get_db)):
+def read_user(user_id: int, db: Session = Depends(database.get_db)):
     db_user = services.get_user(db, user_id=user_id)
     if db_user is None:
         raise HTTPException(status_code=404, detail="User not found")
